mrc-load-model-utils(extract-debug-tensor-dump-hook-move,mechanical_provable): move maybe_register_debug_tensor_dump_hook to load_model_utils module (cut+paste)
PR-Title: Extract load-model helpers into load_model_utils
